Decompose ChatMessage.tsx into thinner composer + sibling files
Issue #1446 calls out ChatMessage.tsx (2,132 lines) as oversized. Four
extractions land the main file at 274 lines (-87%):

- ChatMessage.styles.ts (new): all 46 styled-components plus the
  typeColor helper, re-exported by name. No cross-component selectors
  exist, so styles split cleanly.
- ChatMessageSourcePreview.tsx (new): the source-list popup and per-
  source chip/annotate menu (SourcePreview + SourceItem).
- ChatMessageTimeline.tsx (new): the bordered TimelinePreview panel,
  the per-step CollapsibleTimelineItem, the streaming-state
  StreamingThoughtTicker, and the getTimelineIcon/getTimelineTitle
  helpers shared between them.
- ChatMessageToolUsage.tsx (new): the assistant tool-usage badge and
  hover/keyboard popover (ToolUsageIndicator), together with the
  extractToolCalls and formatToolName helpers it depends on.

Public API (ChatMessage component, ChatMessageProps, TimelineEntry,
StreamingThoughtTicker, extractToolCalls, formatToolName) stays
importable from widgets/chat/ChatMessage via re-exports, so existing
importers (CorpusChat, ChatTray, ApprovalOverlay, ChatSourceAtom,
chat/types, and the unit + component tests) need no changes.

State doesn't need to be lifted to a Jotai atom for this slice; the
extracted components own only local UI state and either receive their
data via props (timeline, sources, availableLabels, createAnnotation)
or already read directly from the existing chatSourcesAtom.

Test plan:
- tsc --noEmit clean
- prettier --check clean
- 12 unit tests in chatMessageUtils.test.ts pass without modification
- 37 component tests across ChatMessageCoverage.ct.tsx,
  ToolUsagePopover.ct.tsx, and chat-message-processing-indicator.ct.tsx
  pass without modification

Refs #1446
